// src/types/generated/inputTypeSchemas/ThumbsignalMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const ThumbsignalMinOrderByAggregateInputSchema: z.ZodType<Prisma.ThumbsignalMinOrderByAggregateInput> = z.object({
  reviewedRestaurantId: z.lazy(() => SortOrderSchema).optional(),
  reviewCreatedBy: z.lazy(() => SortOrderSchema).optional(),
  signalVariant: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional(),
  createdBy: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default ThumbsignalMinOrderByAggregateInputSchema;